sdk/node: read TLS CA from environment
This change is intended to make it easier to write one
Node program that works both on production and inside
with-ledgerd. It adds one new environment variable to
control the Node SDK's behavior:

- SEQTLSCA specifies a path to a CA certificate file.

This removes the agent configuration param on the client
constructor. We are approaching a new major version (v2)
anyway, so that's okay.

Closes #3123

Author: Keith Rarick <kr@xph.us>
Date: Thu Mar 29 16:01:25 2018 -0700
upstream:a88f62294e39aa016106df3ee07a8761ee7e1910